Add parser for data update requests
Add parser that goes through the data.update request json and a map of
sensors/type in order to create the correct KNoTValue instance for
each (sensorId, value) pair.

The data.update request establishes that the value that is being passed
to the thing can belong to any of the supported types. As such there was
no way to determine which object the received value should be casted to.